+++ http/src/test/java/com/twitter/finatra/http/tests/integration/doeverything/main/DoEverythingJavaController.java
@@ -1,4 +1,4 @@
-package com.twitter.finatra.integration.main;
+package com.twitter.finatra.http.tests.integration.main;
 
 import javax.inject.Inject;