+++ http/src/test/java/com/twitter/finatra/http/tests/integration/doeverything/main/HelloService.java
@@ -1,4 +1,4 @@
-package com.twitter.finatra.integration.main;
+package com.twitter.finatra.http.tests.integration.main;
 
 import java.util.HashMap;
 import java.util.Map;