+++ http/src/test/java/com/twitter/finatra/http/tests/integration/doeverything/test/DoEverythingJavaServerFeatureTest.java
@@ -1,4 +1,4 @@
-package com.twitter.finatra.integration.test;
+package com.twitter.finatra.http.tests.integration.doeverything.test;
 
 import org.junit.Assert;
 import org.junit.Test;
@@ -9,9 +9,8 @@ import com.twitter.finagle.http.Request;
 import com.twitter.finagle.http.Response;
 import com.twitter.finagle.http.Status;
 import com.twitter.finatra.http.EmbeddedHttpServer;
+import com.twitter.finatra.http.tests.integration.main.DoEverythingJavaServer;
 import com.twitter.finatra.httpclient.RequestBuilder;
-
-import com.twitter.finatra.integration.main.DoEverythingJavaServer;
 
 public class DoEverythingJavaServerFeatureTest extends Assert {