+++ CHANGELOG.md
@@ -8,6 +8,8 @@ All notable changes to this project will be documented in this file. Note that `
 
 ### Changed
 
+* finatra-http: Move integration tests to a package under `com.twitter.finatra.http`. ``RB_ID=866487``
+
 ### Fixed
 
 ### Closed